SAK-13584
Add integer versions of the top and bottom message numbers to the context.
This allows the option for the tool's markup to produce REST-like URLs
when moving from a list-view to an individual view.  These values are simply
placed in context - it is up to the tool whether it uses them.