SAK-13584
Update the MailArchive 2.6 conversion script.

As part of a 2.6 upgrade this script needs to be run on the database.
It adds a BODY and SUBJECT column and extracts these 
out of the XML to improve searching, sorting, and 
paging performance.

The conversion script is in the mialarchive directory and 
it is called as follows:

./mailarchive-runconversion.sh upgradeschema-2.6-oracle.config 

You must do this with CATALINA_HOME pointing to your Sakai
deployment - it looks through sakai.properties for conneciton 
details and finds the DB drivers in your Tomcat as well.

The convrsion can be run more than once - it notices when a 
message alreayd has been converted and skipps it.  You 
can even drop the columns and then re-run the conversion.

Please put this in 2.6 - it is a blocker.

Note - this is a required conversion as part of 2.6 - it needs to go 
into the release notes somewhere.