SAK-14581 In a clustered environement every server can see all sessions that exist across an entire cluster.  However, for performance reasons we do not want the data for all sessions to EXIST on every server.  Only the data for sessions that server is currently actively using.
Previous to this commit, during ServerComponent$Maintenance.run() the maintenance thread, queries every session object to see if it is expired.  This would have had the negative effect of bringing the session object into memory for every session.  This could potentially have then brought in the attributes of that session, and then one server now sees all data.

To prevent this behavior, ServerComponent$Maintenance.run() was changed, to query a parallel structure which only contains server id and expiration times.  This expiration time is only a suggested possible expiration time.  If it is reached, then the real session object is queried, and asked if it's invalid.  If it is, then behavior is the same as it was before.

To prevent a large set of updates to the parallel structure in the case of a busy/active cluster, the parallel structure is only updated if a session's expiration time is within the second half of being expired.  For example, if expiration is set to 30 minutes, and between minutes 0-15 no update to the parallel structure will take place.  If the expiration time is within minute 16-30 of the expiration window, then the parallel structure will be updated.

The tc-config.xml file was updated to include the apache commons MutableLong as an instrumented class.  This class is used to store the expiration time stamp value.