SAK-2887
Put the slide materials into an iFrame instead of an img tag.

This makes copyright materials work, pdfs work, etc.

Sometimes the browser gets a little confused when it gets redirected from a jpg
to a form - the workaround seems to be for viewers to preview the copyrighted
slides once before the show starts.

But it is much better.