SAK-14449 SAK-14581 SAK-14494
3 different changes.

First (additional root) was for SAK-14581 work.  It was a bug that his was missing from the original check-in for this file for that issue.

Second change (removal of several locks sections) was the verification these lock sections are no longer needed given the other work in SAK-14449.

Third (the removal of the java.util.regex.PatternSyntaxException) was related to the work in SAK-14494.  The line removed did not seem to be necessary, as it was included by accident and is not really an innerclass of Pattern.