SAK-14675 Add support for configuring Terracotta using Terracotta Integration Module (TIM) The terracotta-config module will now deploy the terracotta tc-config.xml and the corresponding TIM jar files, only if explicitly declared to do so by the person running the build.  This can be done by using the following maven command line syntax:
mvn -Dterracotta.enable=true install

Or by setting an active profile element in your settings.xml, e.g.,

   <activeProfiles>
      <activeProfile>terracotta-config</activeProfile>
   </activeProfiles>

It is important to understand the files are copied to the ${maven.tomcat.home}/sakai/tcotta directory, during the 'install' lifecycle and NOT during the sakai:deploy goal like most other things.  This is why it must be explicitly enable and run by the end user.