Back out the change made in r53370 for SAK-14649 for the update() routine.
Still need to use the session variables to differentiate whether this is a realm.upd caused by provider change (the auto adjusting of affiliates are needed) or by user manually remove affiliate member from participant list (then NO further auto adjusting of affiliates are needed)

Added this time is to always convert the event reference to containing site object. Sometimes, the group reference is passed in. It is really the site's url that is needed in the following operations, like the adjusting affiliates, and update umiac provider urls, etc.