SAK-14378
Patch from Savitha 

The Admin realms tool should allow admins to set the active/inactive status of a user. Currently there's no way to do that if the property is set which hides that feature from Site Info. Even if an institution doesn't want to allow instructors to inactivate users, there are times when a provided user is in the site in error and it would be useful to support to be able to inactivate them.

Thanks