NOJIRA. This effectively back-ports the "KNL-76.diff" patch attached to (of course) KNL-76 as of Nov 21, 2008 into the Wiley 2-5-x codebase (cascaded site reference alias deletion triggered by site deletion).
Rather than branch the site module, I elected to keep the new 
NotificationAction in its originally proposed location in the 
site-manage-impl module (but the class name matches that in the 
KNL-76.diff). Since Wiley always deploys the site-manage module, we 
don't have to worry about site-manage-impl not being deployed and 
aliases thus being orphaned on site deletion. So we can tolerate the 
architectural impurity of this factoring until we upgrade to a kernel 
1.1-based release (assuming KNL-76 is accepted into kernel 1.1 of 
course). 

I elected not to back-port the integration test because there is no good 
pre-existing location for such a thing and it is so simple and, frankly, 
unlikely to be frequently executed if at all. Adding the corresponding 
machinery just is not worth it.

(EDUGEN-544)