SAK-15097
This adds support for a new site property called sakai:htmlInclude - when 
this property is present, it is included in the site navigation just below
the icon in a div titled "siteHTMLInclude" - if this property is not set 
on a site nothing is changed.  The typical use case for this might be to add
a bit of Javascript to support a tool such as hab.la to a site.   Since
this is off by default and only the admin user can set this property, this
feature will have no impact unless the admin user sets the property on 
the site.