BSP-1917 - Added Instructor to allowedroles for Head GSI.  Data for allowed role filter now contains 				<entry> 					<key> 						<value>Head GSI</value> 					</key> 					<set> 						<value>Instructor</value> 						<value>Head GSI</value> 						<value>GSI</value> 						<value>Librarian</value> 						<value>Dept. Delegate</value> 						<value>Student</value> 						<value>Guest</value> 					</set> 				</entry>
This means that a Head GSI but NOT a GSI can promote any user to Instructor.  Fixes problem where Instructors appeared to be Dept. Delegates on SiteInfo landing page.  This was due to the fact that Instructor was not an Option in the select list and the visible select option, Dept. Delegate, was shown just because it is first in the alphabetically sorted list.