KNL-218 Only synchronize request threads on session when in a Terracotta cluster.
This may be problematic for Terracotta and needs a better longer-term solution, otherwise Terracotta clusters are likely to suffer from performance problems with user threads being blocked, especially in multi-tab browsing.

This change is to restore the pre-Terracotta behaviour when Terracotta is not enabled. There are still 2 other uses of synchronized in RequestFilter that have been introduced but they're less likely to cause problems.