BSP-2119 - existing code always reads the roster.defaultSortColumn from sakai.properties or default.sakai.properties if it exists, as it does in default.sakai.properties in branch 2.6_x BUT NOT 2.5_x(where it is commented out in the component sakai.properties), and never reads the current state of the sortColumn in RosterPreferences.  In that case, it always sorts by displayName no matter which column heading is clicked on.  If the property does NOT exist, code defaults to dispayName only (is not configurable) and DOES respond to the current state of sortColumn.
Basically, the code partially works if property does not exist as in 2.5_X (but is not configurable) and breaks if configuration property does exist as in 2.6_x and UC Berkeley's current msub version.

Fixed by moving code that reads the sakai.properties value to RosterPreferences class itself and wiring prefs bean to roster's Services class.  Code will now read any valid existing roster.defaultSortColumn property and use it for the initial page load sort order.   If property is missing or has invalid value, it will default to displayName for initial sort order and log.WARN about missing or invalid property.