SAK-16450 -  QA at UCB caught an oversight where enrollment status sorting was broken. caused by the fact that the same class RosterPreferences was being used by two managed beans, "prefs" and "enrollmentStatusPrefs" and I did not realize the second was in play and did not make the necessary change in enrollmentStatusPrefs that I had made in prefs in the JSF config file so that in enrollmentStatusPrefs the services reference was null.
Overall, the existing code always reads the roster.defaultSortColumn from sakai.properties or default.sakai.properties if property exists, as it does in default.sakai.properties in branch 2.6_x BUT NOT 2.5_x(where it is commented out in the component sakai.properties), and never reads the current state of the sortColumn in RosterPreferences. In that case, it always sorts by displayName no matter which column heading is clicked on. If the property does NOT exist, code defaults to dispayName only (is not configurable) and DOES respond to the current state of sortColumn.

Basically, the code partially works if property does not exist as in 2.5_X (but is not configurable) and breaks if configuration property does exist as in 2.6_x and UC Berkeley's msub 2.6_x version.

Fixed by moving code that reads the sakai.properties value to RosterPreferences class itself and wiring prefs bean to roster's Services class. Code will now read any valid existing roster.defaultSortColumn property and use it for the initial page load sort order. If property is missing or has invalid value, it will default to displayName for initial sort order and log.WARN about missing or invalid property.

For QA testing -
Expected Valid Results
1) in the runtime sakai.properties uncomment one of the following properties at a time and restart the server after changing the property:
#roster.defaultSortColumn=sortName
#roster.defaultSortColumn=role
#roster.defaultSortColumn=email
#roster.defaultSortColumn=displayId

Select roster tool and confirm that the initial sort column corresponds to the uncommented property. Sort order should be ascending and show the proper sort icon.  
Confirm that clicking on any other column will initially sort on the clicked column in ascending order.
Subsequent clicks on the same column will reverse(toggle) the sorting order

Expected Invalid Results
2) Enter an invalid sort Column in the runtime sakai.properties - e.g roster.defaultSortColumn=invalidName
Select roster tool and confirm that the initial sort column is "sortName".  Confirm also the appearance of the following log.WARN statements like those below:
2009-09-28 18:32:46,133  WARN http-8080-Processor25 org.sakaiproject.tool.roster.RosterPreferences - defaultSortColumn value = invalidName is invalid, must be one of [sortName, email, role, displayId]
2009-09-28 18:32:46,133  WARN http-8080-Processor25 org.sakaiproject.tool.roster.RosterPreferences - Check your default.sakai.preferences and sakai.preferences for a valid roster.defaultSortColumn value
2009-09-28 18:32:46,133  WARN http-8080-Processor25 org.sakaiproject.tool.roster.RosterPreferences - Defaulting to sorting by sortName

Also, please confirm that sorting on the Enrollment Status page is working

Please backport these changes to the 2.5.x