SAK-17598 - Allow free-form portfolio presentations to be disabled
Implemented as described on the ticket. There is a new property, osp.freeform.disabled, which can be applied in sakai.properties or as a site property. When specified as a site property, it takes precedence. When specifying the site specific setting, the strings "true" or "false" should be used. The default is false, which results in no behavioral change.