SAK-2991
Add a Cancel option when there is somewhere to cancel to.  SO when the 
login is coming from a Portal login, the user gets a cancel option.  However
when they are coming from a bookmarked site or a /access URL they will not get
a cancel option because there is nowhere to "go back to".