SAK-18864
Abstracted the EventManager such that the old implementation can still be used if one prefers. Then created a Hibernate-backed version.
Added a pager to the events log, then filtering to filter on date, type, or job name.
Added other types of events and a message property to events in anticipation of adding the ability to log from jobs to the log