SAK-19178
SchedulerManagerImpl is now wired with a set of SpringInitialJobSchedules from the application context at startup. A test is run to determine if the Quartz database tables have been created yet. If not it is assumed that this is the first time Sakai has run and the SpringInitialJobSchedules are used to create jobs and triggers as configured.
This behavior is controlled by a server configuration parameter scheduler.loadjobs. It can be one of three values:

	init - (default behavior) schedule jobs on the first run of Sakai
	true - schedule jobs every time Sakai restarts
	false - never schedule the jobs

This mechanism is used to schedule the Event Log Purge job to run nightly at midnight.