SAK-19152
This changes the site.title.maxlength default to be 25 so 
"Administration Workspace" is not chopped out of the box.

If site.title.maxlength>=10, when a title is truncated, 
we add " …" to the title so if the length were 10, a 
title of "ReallyLongTitle" would become "ReallyL …" 
so even with the … the overall title will still be 
10 characters.  If site.title.maxlength<10 we simply 
truncate the title to the specified length.

This is now changed both in site-manage and in 
the portal to be the same (25).