SAK-19455
Allow for a special portlet (the Neo Portlet) to affect the portal navigation

- Set portal breadcrumb navigation allowing navigation into its own internal hierarchy
- Optionally suppressing their own title bar
- Optionally handling their own reset, help, and edit mode buttons - the URLs for these will be provided to the portlet so it can display them in its own UI

Sample code for a portlet that accomplishes this is here:

https://source.sakaiproject.org/contrib/csev/trunk/neo/neo-portlet

This is controlled by a property - the portal can prohibit Neo portlets by adding:

portal.allow.neo.portlet=false

At this point, there are no Neo portlets.  My hope is that Lesson Builder may reconstitute
itself as a Neo portlet.