SAK-19496
Make sure that when a JSR-168 portlet makes an honest mistake 
and coughs up an Exception, that the portal displays it in
context rather than making it look like portal tossed its
cookies.

This would be nice to back-port to 2-7-x and 2-8-x at some 
point - but it may *appear* to conflict with r86609 - it
does not conflict - but the lines in SkinnableCharonPortal.java
around the result.getContent(); line were changed in r86609
so a patcher might get concerned.  Just revert r86609 and then 
make the patch and all should be nice and clean.