SAK-20463 SAK-19813
Makes it so that sakai_action parameters are passed from portal urls into the
tool/iframe urls displayed in the ultimate page.

Restructure the nav to allow multiple tools/shortcuts in a drop down instead 
of a single preferences link (i.e. the gear icon).

Thanks to Gonzalo for the extreme awesomeness of the UI.