SAK-20712
Allow legacy tools to have neo-lite like behavior.   This allows a legacy tool to request
that the portal suppredd the display of the title bar, help, and reset button.   The tool
output is tucked right up againse the portal navigation.

The tool indicates it wantes this behavior by setting a tool placement property:

sakai-portal:suppresstitle=true 

The tool will be handed the reset and help urls in its tool session if it wants to display these 
in its own UI.

sakai-portal:help-action 
sakai-portal:reset-action 

Help should be a new window - reset should replace the current iframe.

Tools can also ask for this as default at tool registration time with this:

<configuration name="sakai-portal:suppresstitle" value="true" />

This is aimed at full-window tools like LessonBuilder, Melete, LOI, OpenSyllabus, etc that
themselves have complex navigation internally.    Tools should be careful to style themselves
so that they look nice with or without the title bars because the system 
admins can revoke a tool's ability to hide the title.   If a tool does not see a reset url
in the session variable sakai-portal:reset-action - it means that the portal/admin has decided
to give the tool a title - so generate a UI accordingly.