SAK-20799
Add portal support for a new property on a tool placement:

sakai-portal:visible=false/true

This is similar to adding site.upd to the functions.require property except that it only
hides the tool from the left nav.   If add a functions.require to have site.upd, the portal
blocks all access to the tool through /portal/tool/placement-id

The hiding of a tool with this new property is only cosmetic and /portal/tool URLs 
continue to work.

If you have site.upd, tools that are hidden or disabled by site.upd are shown to site
maintainers in a different font.  If you don't have site.upd the pages are not there.