SAK-20799
Add page order support for a new property on a tool placement:

sakai-portal:visible=false/true

Now instead of a show/hide that is a lightbulb on or off, we also have an
enable/disable button.

Show/hide now is only a cosmetic change that hides the tool from users w/o
site.upd and shows the link in a different font for those with site.upd.

Enable/disable (with a lock icon) now changes the functions.required to include
site.upd (like the ond light bulb) - this hides the tool from non site.upd
users and blocks access to the tool via /portal/tool.  For those with site.upd
the tool functions but is shown in a different font.

There is no conversion needed, the pre-2.9 behavior of show/hide is exactly
the same as the new enable/disable button.  If a user hid a tool in pre 2.9,
it will appear post-2.9 as *both* hidden and disabled with no conversion 
necessary as the underlyind aya model is unchanged.  The only post-2.9 
difference is that they will have finer-grain control over these functions.

Decoupling Hide/Show from Enable/Disable is necessary for teachers that want to 
use Lesson Builder to orchestrate Samigo, Forums, etc but do not want to show
those tools in the left nav.   The former behavior only gave them a choice to hide
the tool and break the tool and the tool would appear as broken when LessonBuilder
tried to launch the tool.  Now they can hide the tool without "breaking" the ability
for Lesson Builder to use the tool.