ONC-3990
When instructors would create an assessment with a start or due date, it was possible to have an exception without a good error message because there was an improper test case for time portion of the date. It was set to remeasure the entire date array's length, rather than checking the time array portion itself.