SAK-20774
Add Basic LTI Administration tool.  This tool is intended to be added
to the !admin site and is an extention to Site Info.

It allows administrators and instructors to make Basic LTI Tools and
Content Items.   Each tool can hae menay content items.   Content Items
have Launch URLs that can be used anywhere within Sakai like Resources,
the iFrame Tool, Melete, literally anywhere.

In particular, this lays a tool and service foundation for Lesson
builder to easily add an "Add External Tool" feature both for
its authoring and when it is importing IMS Common Cartridges 
(version 1.1).

This also lays a foundation / starting point to add support for
auto-provisioning capabilities for Full LTI.

This design is informed by the great work Basic LTI work in Moodle,
ATutor, Blackboard, Desire2Learn, and Instructure's Canvas.  As more 
experience is gained in the use of BLTI in LMS systems, the UIs are
starting to converge.