svn merge -c95423 https://source.sakaiproject.org/svn/msub/iu.edu/oncourse/trunk/sam --- Merging r95423 into 'samigo-app': U    samigo-app/src/java/org/sakaiproject/tool/assessment/ui/bean/author/AssessmentSettingsBean.java svn log -c95423 https://source.sakaiproject.org/svn/msub/iu.edu/oncourse/trunk/sam ------------------------------------------------------------------------ r95423 | cajrusse@indiana.edu | 2011-07-20 11:38:34 -0400 (Wed, 20 Jul 2011) | 3 lines
ONC-3990

When instructors would create an assessment with a start or due date, it was possible to have an exception without a good error message because there was an improper test case for time portion of the date. It was set to remeasure the entire date array's length, rather than checking the time array portion itself.
------------------------------------------------------------------------