SAK-21126 - Changing Metaobj XSLT binding to rich text editor for better backward compatibility.
The original 2.8.0 change was to replace the result of
XsltFunctions.getRichTextScript with a simple JavaScript call to
sakai.editor.launch. Because of the proliferation of custom creation
renderer XSLT files, this turns out to be a bad idea. This change
replaces the original result of getRichTextScript and adds a new method
that can be migrated to. See the comments in source for details.