BLTI-121
Initial commit of the left hand navigation addition (aka 
Leftify).  This contains a simple replacement for the 
web/iframe portlet (under a different tool id) that is 
based on JSR-168 so we don't end up with iframe-within-iframe 
issues.

The JSR-168 iframe portlet will be hidden after testing is
done as its only intent is to be used as the way to show 
a tool configuration in the left nav for now.