SAK-13821 Use the maven antrun plugin to concatenatie files.
The issue is that the raw default.sakai.propertis file is pulled by developers and deployers from SVN and used to configure their instances. Having unexpanded properties in this file causes Sakai to fail to start, so moving the automatically expanded properties to another file and putting the two files back together in the build keeps everyone happy.
There doesn't seem to be a nicer way to concat files when using maven than using the antrun plugin.