BLTI-129
Finish the stealth feature.  Allows admin to install a tool marking it 
as stealth and make it available to all sites.  It will not be visible
to instructors in their tool list.  The Admin can push a tool configuration
into a site from the Admin External Tools tool (i.e. not from Site Info).

Once a stealthed tool configuration is made available to a site, the 
Instructor can edit the editable fields, test, and even delete the 
tool configuration.  But they cannot create a new tool configuration
since they cannot see the installed tool.