SAM-1438 - Adjusting DeliveryBean settings getter to return safely
This is a workaround to try to avoid the null bean issue and at the same time
produce some diagnostic information about conditions that cause the issue. The
temporary settings object is populated from the assessment if available, or
returned as the defaults if not, such that getSettings() never returns null and
takes a best effort at returning the right information without changing the
member variable.