KNL-1007 Add support for HTML5 data-* and valueless attributes
This change is local to FormattedText except for the addition to the allowed
attributes in config. It changes the parsing/checking of attributes to be more
thorough and allows two constructs that were previously rejected: HTML5 data-*
attributes and valueless attributes. The API is unchanged.

There is a new test added to test new and existing attribute patterns. All
existing tests are passing and editing with CKEditor behaves the same in basic
testing.