BLTI-207 BLTI-203
Cleanup a few hanging chads w.r.t. BLTI-203

Teach the Site Info tool to be able to set the allowlori field on its tools.

Catch a few printlns and turn them into log.debug()

Catch a little bug where it was trying to find the pre-installed tool and making 
the wrong if-test.