LSNBLDR-236 - Update CLOB columns from text to materialized_clob
A few columns were recently changed to use the "text" type and, since
the Hibernate 3.6 upgrade, these should use "materialized_clob". This
is because "text" maps to LONG on Oracle, which is strongly
discouraged. The CLOB type is better behaved in all respects as far as
I am aware. The new Hibernate type was added to relieve the ambiguity
of coercing types on a combination of mapped type and length limit.

See also: KNL-515 and https://community.jboss.org/wiki/HibernateCoreMigrationGuide35