Correct fix for r132949
I'd assumed that the kaltura plugin was being loaded before CKEditor, but that
wasn't right.  Introducing the $(document).ready fixed the issue (but not for
the reason I thought it did), but created problems of its own when the RTE was
loaded in a context where JQuery wasn't available.

The real problem was this line:

  var CKEDITOR = CKEDITOR || {};

Was actually causing the CKEDITOR global to be reset in IE10.  I guess the "var
CKEDITOR" was being hoisted and causing CKEDITOR to be set to 'undefined', even
in cases where it might have previously had a value.